+++ mutations/ve_b_outside_plan_guard.pl
@@ -1,0 +1,1 @@
+s/if \(!\(\$inplan \?\? helper::competency_in_plan\(\$competencyid, \$plan\)\)\) \{\s*return false;\s*\}\s*//;